+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 ## Next version
 
+## 0.34.0
+
 ### Added
 
 * Version command https://github.com/xcbuddy/xcbuddy/pull/25 by @pepibumur.